Use unsigned integers within workflow attributes (#243)

+++ planetscale/workflows.go
@@ -12,7 +12,7 @@ import (
 type Workflow struct {
 	ID                   string     `json:"id"`
 	Name                 string     `json:"name"`
-	Number               int        `json:"number"`
+	Number               uint64     `json:"number"`
 	State                string     `json:"state"`
 	CreatedAt            time.Time  `json:"created_at"`
 	UpdatedAt            time.Time  `json:"updated_at"`
@@ -80,9 +80,9 @@ type WorkflowTable struct {
 	Name           string    `json:"name"`
 	CreatedAt      time.Time `json:"created_at"`
 	UpdatedAt      time.Time `json:"updated_at"`
-	RowsCopied     int64     `json:"rows_copied"`
-	RowsTotal      int64     `json:"rows_total"`
-	RowsPercentage int       `json:"rows_percentage"`
+	RowsCopied     uint64    `json:"rows_copied"`
+	RowsTotal      uint64    `json:"rows_total"`
+	RowsPercentage uint      `json:"rows_percentage"`
 }
 
 type WorkflowVDiff struct {
@@ -93,8 +93,8 @@ type WorkflowVDiff struct {
 	StartedAt          *time.Time                 `json:"started_at"`
 	CompletedAt        *time.Time                 `json:"completed_at"`
 	HasMismatch        bool                       `json:"has_mismatch"`
-	ProgressPercentage int                        `json:"progress_percentage"`
-	EtaSeconds         int64                      `json:"eta_seconds"`
+	ProgressPercentage uint                       `json:"progress_percentage"`
+	EtaSeconds         uint64                     `json:"eta_seconds"`
 	TableReports       []WorkflowVDiffTableReport `json:"table_reports"`
 }
 
@@ -123,7 +123,7 @@ type ListWorkflowsRequest struct {
 type GetWorkflowRequest struct {
 	Organization   string `json:"-"`
 	Database       string `json:"-"`
-	WorkflowNumber int    `json:"-"`
+	WorkflowNumber uint64 `json:"-"`
 }
 
 // WorkflowsService is an interface for interacting with the workflow endpoints of the PlanetScale API
@@ -180,6 +180,6 @@ func workflowsAPIPath(org, db string) string {
 	return fmt.Sprintf("%s/%s/workflows", databasesAPIPath(org), db)
 }
 
-func workflowAPIPath(org, db string, number int) string {
+func workflowAPIPath(org, db string, number uint64) string {
 	return fmt.Sprintf("%s/%d", workflowsAPIPath(org, db), number)
 }
